Refactor structs and enums for improved clarity
- Updated `CsvReportedError`, `CsvError`, `ValidationIssue`,
  `ValidationResult`, and `ValidationRule` structs to use
  expression-bodied constructors for automatic property
  initialization, simplifying the code.

- Assigned explicit integer values to `CsvErrorType` and
  `ValidationSeverity` enums for better clarity and control.

- Changed `RecordCount` property type in `ICsvReader`
  from `long` to `int` to limit the number of records processed.

+++ src/FastCsv/ICsvReader.cs
@@ -19,7 +19,7 @@ public partial interface ICsvReader
     /// <summary>
     /// Total number of records processed so far
     /// </summary>
-    long RecordCount { get; }
+    int RecordCount { get; }
 
     /// <summary>
     /// Current position in the input data
